Moves Card, LinkCard and MediaCard to components folder
