+++ src/components/cards/LinkCardB.tsx
@@ -1,8 +1,8 @@
 import React from "react";
-import { Content } from "../../../../api";
-import { TableRowCell } from "../../../../layout/Table";
-import { TdCSS, LinkCSS, ImageCSS } from "../../../../css";
-import { headline } from "../../../../styles/typography";
+import { Content } from "../../api";
+import { TableRowCell } from "../../layout/Table";
+import { TdCSS, LinkCSS, ImageCSS } from "../../css";
+import { headline } from "../../styles/typography";
 
 type Theme = "light" | "dark";
 
@@ -25,7 +25,7 @@ const iconStyle: ImageCSS = {
     border: "0"
 };
 
-export const LinkCard: React.FC<{
+export const LinkCardB: React.FC<{
     content: Content;
     theme: Theme;
 }> = ({ content, theme }) => {
