+++ src/components/cards/MediaCardC.tsx
@@ -1,10 +1,9 @@
 import React from "react";
-import { FontCSS, TdCSS, TableCSS, ImageCSS } from "../../../../css";
+import { TdCSS, ImageCSS } from "../../css";
 import { palette } from "@guardian/src-foundations";
-import { Content } from "../../../../api";
-import { formatImage } from "../../../../image";
-import { Table, RowCell, TableRowCell } from "../../../../layout/Table";
-import { Padding } from "../../../../layout/Padding";
+import { Content } from "../../api";
+import { formatImage } from "../../image";
+import { Table, RowCell, TableRowCell } from "../../layout/Table";
 
 const tdStyle: TdCSS = {
     backgroundColor: palette.neutral[100],
@@ -57,7 +56,7 @@ const Image: React.FC<{
     );
 };
 
-export const MediaCard: React.FC<Props> = ({ content, salt }) => {
+export const MediaCardC: React.FC<Props> = ({ content, salt }) => {
     const image =
         content.properties.maybeContent.trail.trailPicture.allImages[0];
     const formattedImage = formatImage(
