+++ src/fronts/opinion/variantB/components/CommentCollection.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { Collection as ICollection } from "../../../../api";
 import { Heading } from "./../../../../components/Heading";
 import { Multiline } from "./../../../../components/Multiline";
-import { Card as CommentCard } from "./Card";
+import { CommentCardB } from "../../../../components/cards/CommentCardB";
 import { Padding } from "../../../../layout/Padding";
 import { Grid as CommentGrid } from "./Grid";
 
@@ -33,7 +33,7 @@ export const CommentCollection: React.FC<{
             <Multiline topPadding />
             <Heading heading={collection.displayName} />
 
-            <CommentCard
+            <CommentCardB
                 content={c0}
                 salt={salt}
                 size={"large"}
@@ -41,7 +41,7 @@ export const CommentCollection: React.FC<{
             />
             <Padding px={12} />
 
-            <CommentCard
+            <CommentCardB
                 content={c1}
                 salt={salt}
                 size={"large"}
@@ -57,7 +57,7 @@ export const CommentCollection: React.FC<{
                 )}
             />
 
-            <CommentCard
+            <CommentCardB
                 content={c6}
                 salt={salt}
                 size={"large"}
@@ -73,7 +73,7 @@ export const CommentCollection: React.FC<{
                 )}
             />
 
-            <CommentCard
+            <CommentCardB
                 content={c9}
                 salt={salt}
                 size={"large"}
