+++ src/fronts/opinion/variantC/components/Grid.tsx
@@ -1,13 +1,16 @@
 import React from "react";
-import { ContributorImageWrapper, getContributor } from "./Card";
 import { palette } from "@guardian/src-foundations";
 import { TableRow, TableRowCell } from "../../../../layout/Table";
 import { Padding } from "../../../../layout/Padding";
 import { Content } from "../../../../api";
 import { GridRow, partition } from "../../../../layout/Grid";
 import { TableCSS } from "../../../../css";
-import { Card as CommentCard } from "./Card";
-import { LinkCard } from "./LinkCard";
+import {
+    CommentCardC,
+    ContributorImageWrapper,
+    getContributor
+} from "../../../../components/cards/CommentCardC";
+import { LinkCardC } from "../../../../components/cards/LinkCardC";
 
 interface CommentGridProps {
     content: Content[];
@@ -50,7 +53,7 @@ export const Grid: React.FC<CommentGridProps> = ({
             <React.Fragment key={i}>
                 <GridRow
                     left={
-                        <CommentCard
+                        <CommentCardC
                             content={pair[0]}
                             salt={salt}
                             size={"small"}
@@ -58,7 +61,7 @@ export const Grid: React.FC<CommentGridProps> = ({
                         />
                     }
                     right={
-                        <CommentCard
+                        <CommentCardC
                             content={pair[1]}
                             salt={salt}
                             size={"small"}
@@ -131,8 +134,8 @@ export const LinkGrid: React.FC<LinkGridProps> = ({ content, salt }) => {
     const rows = partition(content, 2).map((pair, i) => (
         <React.Fragment key={i}>
             <GridRow
-                left={<LinkCard content={pair[0]} />}
-                right={pair[1] ? <LinkCard content={pair[1]} /> : null}
+                left={<LinkCardC content={pair[0]} />}
+                right={pair[1] ? <LinkCardC content={pair[1]} /> : null}
                 leftStyles={{
                     backgroundColor: palette.neutral[100],
                     borderLeft: `1px solid ${palette.neutral[20]}`,
