+++ src/fronts/opinion/variantC/components/MediaCollection.tsx
@@ -2,14 +2,14 @@ import React from "react";
 import { Collection as ICollection } from "../../../../api";
 import { Heading } from "./../../../../components/Heading";
 import { Multiline } from "./../../../../components/Multiline";
-import { MediaCard } from "./MediaCard";
+import { MediaCardC } from "../../../../components/cards/MediaCardC";
 
 export const MediaCollection: React.FC<{
     collection: ICollection;
     salt: string;
 }> = ({ collection, salt }) => {
     const items = collection.backfill.map(content => (
-        <MediaCard content={content} salt={salt} />
+        <MediaCardC content={content} salt={salt} />
     ));
 
     return (
